refactor: move request and response into new files
