+++ src/handler.rs
@@ -4,6 +4,8 @@ use async_trait::async_trait;
 use futures::Future;
 use serde::{de::DeserializeOwned, Serialize};
 
+use crate::{error::Error, request::Request, response::Response};
+
 type StandardBodyType = String;
 pub type GenericRequest = Request<StandardBodyType>;
 pub type GenericResponse = Response<StandardBodyType>;
@@ -12,131 +14,7 @@ pub type BoxedHandler =
 pub type RefHandler<'a> =
     &'a dyn Fn(GenericRequest) -> Pin<Box<dyn Future<Output = GenericResponse>>>;
 
-#[derive(Debug)]
-pub struct Error {
-    body: String,
-}
-
-impl Error {
-    pub fn new(body: String) -> Self {
-        Self { body }
-    }
-}
-
-impl From<Error> for GenericResponse {
-    fn from(val: Error) -> Self {
-        GenericResponse::new(val.body)
-    }
-}
-
-pub type Method = http::Method;
-pub type Uri = http::Uri;
-pub type HttpRequest<T> = http::Request<T>;
-pub type HttpBuilder = http::request::Builder;
-
-pub struct Request<T> {
-    request: HttpRequest<T>,
-}
-
-pub struct Builder {
-    pub builder: HttpBuilder,
-}
-
-impl Builder {
-    pub fn uri<T>(self, uri: T) -> Builder
-    where
-        Uri: TryFrom<T>,
-        <Uri as TryFrom<T>>::Error: Into<http::Error>,
-    {
-        Self {
-            builder: self.builder.uri(uri),
-        }
-    }
-
-    pub fn body<T>(self, body: T) -> Request<T> {
-        Request {
-            request: self.builder.body(body).unwrap(),
-        }
-    }
-
-    pub fn method(self, method: Method) -> Self {
-        Self {
-            builder: self.builder.method(method),
-        }
-    }
-}
-
-type Result<T> = std::result::Result<T, Error>;
-
-impl<T> Request<T> {
-    pub fn new(value: T) -> Self {
-        Self {
-            request: HttpRequest::new(value),
-        }
-    }
-
-    pub fn body(&self) -> &T {
-        self.request.body()
-    }
-
-    // TODO: Valuate if this will keep this fn or move to an from_parts style
-    fn and_then<BodyType>(
-        self,
-        callback: impl FnOnce(T) -> Result<BodyType>,
-    ) -> Result<Request<BodyType>> {
-        let (parts, body) = self.request.into_parts();
-        callback(body).map(|body| Request {
-            request: HttpRequest::from_parts(parts, body),
-        })
-    }
-
-    pub fn method(&self) -> &Method {
-        self.request.method()
-    }
-
-    pub fn method_mut(&mut self) -> &mut Method {
-        self.request.method_mut()
-    }
-
-    pub fn uri(&self) -> &Uri {
-        self.request.uri()
-    }
-
-    pub fn uri_mut(&mut self) -> &mut Uri {
-        self.request.uri_mut()
-    }
-}
-
-impl Request<()> {
-    pub fn builder() -> Builder {
-        Builder {
-            builder: HttpBuilder::new(),
-        }
-    }
-}
-
-pub struct Response<T> {
-    body: T,
-}
-
-impl<T> Response<T> {
-    pub fn new(value: T) -> Self {
-        Self { body: value }
-    }
-
-    pub fn body(&self) -> &T {
-        &self.body
-    }
-
-    // TODO: Valuate if this will keep this fn or move to an from_parts style
-    fn and_then<BodyType>(
-        self,
-        callback: impl FnOnce(T) -> Result<BodyType>,
-    ) -> Result<Response<BodyType>> {
-        let body = self.body;
-        callback(body).map(Response::<BodyType>::new)
-    }
-}
+pub type Result<T> = std::result::Result<T, Error>;
 
 pub trait BodyDeserializer {
     type Item: DeserializeOwned;
@@ -274,9 +152,7 @@ where
     where
         Self: std::marker::Sized,
     {
-        serde_json::from_str(content).map_err(|err| Error {
-            body: err.to_string(),
-        })
+        serde_json::from_str(content).map_err(|err| Error::new(err.to_string()))
     }
 }
 
@@ -287,9 +163,7 @@ where
     type Item = T;
 
     fn serialize(content: &Self::Item) -> Result<String> {
-        serde_json::to_string(content).map_err(|err| Error {
-            body: err.to_string(),
-        })
+        serde_json::to_string(content).map_err(|err| Error::new(err.to_string()))
     }
 }
